+++ CHANGELOG.md
@@ -13,9 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 -   Expose parse\* methods as static methods for Valhalla ([#35](https://github.com/gis-ops/routingjs/pull/35))
 -   This Changelog ;-) ([#35](https://github.com/gis-ops/routingjs/pull/35))
+-   `/trace_route` support for Valhalla [#34](https://github.com/gis-ops/routingjs/pull/34)
+-   Basic waypoint support (mainly for Valhalla) [#40](https://github.com/gis-ops/routingjs/pull/40)
 
 ### Fixed
 
 -   Remove `directionsOpts` as top level options in `ValhallaTraceRouteOpts` ([#35](https://github.com/gis-ops/routingjs/pull/35))
 -   Remove User-Agent header ([#33](https://github.com/gis-ops/routingjs/pull/33))
-